src/services/image_embedding_service/pyproject.toml
[project]
name = "image-embedding-service"
version = "0.1.0"
description = "Transforms images into embeddings using Hugging Face vision models"
authors = [
    { name = "Tamas Szombathy", email = "szombathy.tamasgergely@gmail.com" }
]
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "transformers",
    "torch",
    "Pillow",
    "python-multipart",
    "aiohttp>=3.12.14",
    "pydantic-settings>=2.10.1",
]
requires-python = ">=3.11.0"

